move LobbyTeamView to components
